MDC-104 Flutter:Material Advanced Components (Flutter)
change the UI in the Shrine app to a two-level presentation called a "backdrop". The backdrop includes a menu that lists selectable categories used to filter the products shown in the asymmetrical grid. In this codelab, you'll use the following Flutter components:

Shape
Motion
Flutter widgets (that you've used in the previous codelabs)